+++ gridsome/app/graphql/guard.js
@@ -1,8 +1,14 @@
 import fetch from '../fetch'
+import config from '~/.temp/config'
 import { getResults, setResults, formatError } from './shared'
 
 export default (to, from, next) => {
   if (process.isServer) return next()
+
+  if (to.meta && to.meta.__custom) {
+    global.__INITIAL_STATE__ = null
+    return next()
+  }
 
   if (process.isProduction && global.__INITIAL_STATE__) {
     setResults(to.path, global.__INITIAL_STATE__)
@@ -23,12 +29,12 @@ export default (to, from, next) => {
       }
     })
     .catch(err => {
-      if (err.code === 'MODULE_NOT_FOUND') {
-        console.error(err) // eslint-disable-line
-        next({ name: '*', params: { 0: to.path }})
-      } else if (err.code === 404 && to.path !== window.location.pathname) {
-        // reload page if coming from another page and data doesn't exist
-        window.location.assign(to.fullPath)
+      if (err.code === 'MODULE_NOT_FOUND' || err.code === 404) {
+        console.error(err)
+        next({ name: '*', params: { 0: to.path } })
+      } else if (err.code === 'INVALID_HASH' && to.path !== window.location.pathname) {
+        const fullPathWithPrefix = (config.pathPrefix ?? '') + to.fullPath
+        window.location.assign(fullPathWithPrefix)
       } else {
         formatError(err, to)
         next(err)