+++ gridsome/lib/app/codegen/routes.js
@@ -1,82 +1,120 @@
-const path = require('path')
-const { uniqBy, isPlainObject } = require('lodash')
-const { NOT_FOUND_NAME } = require('../../utils/constants')
+const slash = require('slash')
+const { relative } = require('path')
 const { slugify } = require('../../utils')
+const { pathToFilePath } = require('../../pages/utils')
+const { NOT_FOUND_NAME } = require('../../utils/constants')
+const { uniqBy } = require('lodash')
 
-function genRoutes (app, routeMeta = {}) {
-  let res = ''
+const isUnitTest = process.env.GRIDSOME_TEST === 'unit'
 
-  const pages = uniqBy(app.pages.data(), page => page.route)
-  const notFound = app.pages.findPage({ name: NOT_FOUND_NAME })
+function genRoutes(app) {
+  const redirects = app.config.redirects.filter(rule => rule.status === 301)
+  const fallback = app.pages._routes.findOne({ name: NOT_FOUND_NAME })
+  const components = []
+  const items = []
 
-  // use the /404 page as fallback route
-  pages.push({
-    ...notFound,
-    dataInfo: routeMeta[notFound.route],
-    name: '*',
-    route: '*'
+  const createRouteItem = (route, name = route.name, path = route.path) => ({
+    name,
+    path,
+    chunkName: genChunkName(app.context, route),
+    variableName: components.includes(route.component)
+      ? `c${(components.indexOf(route.component) + 1)}`
+      : `c${components.push(route.component)}`,
+    meta: route.internal.meta,
+    type: route.type,
+    component: isUnitTest
+      ? relative(app.context, route.component)
+      : route.component
   })
 
-  res += `export default [${pages.map(page => {
-    const component = JSON.stringify(page.component)
-    const chunkName = JSON.stringify(page.chunkName || genChunkName(page.component, app.context))
-    const dataInfo = page.dataInfo || routeMeta[page.route]
-    const hasContext = Object.keys(page.context).length > 0
-    const props = []
-    const metas = []
+  for (const redirect of redirects) {
+    items.push(redirect)
+  }
 
-    props.push(`    path: ${JSON.stringify(page.route)}`)
-    props.push(`    component: () => import(/* webpackChunkName: ${chunkName} */ ${component})`)
+  for (const route of app.pages.routes()) {
+    items.push(createRouteItem(route))
+  }
 
-    if (typeof dataInfo === 'string') {
-      metas.push(`data: () => import(/* webpackChunkName: ${chunkName} */ ${JSON.stringify(dataInfo)})`)
-    } else if (Array.isArray(dataInfo)) {
-      metas.push(`data: ${JSON.stringify(dataInfo)}`)
-    } else if (process.env.GRIDSOME_MODE !== 'static' && (page.query.document || hasContext)) {
-      metas.push(`data: true`)
+  // use the /404 page as fallback route
+  if (fallback) {
+    items.push(createRouteItem(fallback, '*', '*'))
+  }
+
+  const routes = items.map(item => {
+    if (item.from && item.to) {
+      return genRedirect(item)
     }
+    return genRoute(item)
+  })
 
-    for (const key in page.internal.meta) {
-      const value = page.internal.meta[key]
+  const componentItems = uniqBy(
+    items.filter(item => item.component),
+    'component'
+  )
 
-      if (isPlainObject(value)) {
-        switch (value.type) {
-          case 'import': {
-            const path = JSON.stringify(value.path)
-            const chunkName = JSON.stringify(value.chunkName || page.chunkName)
-            const importCode = `import(/* webpackChunkName: ${chunkName} */ ${path})`
+  return [
+    `${componentItems.map(genComponent).join('\n')}\n\n`,
+    `export default [${routes.join(',')}\n]\n`
+  ].join('')
+}
 
-            metas.push(`${key}: () => ${importCode}`)
+function genComponent(item) {
+  const component = JSON.stringify(item.component)
+  const chunkName = JSON.stringify(item.chunkName)
 
-            break
-          }
-          case 'code': {
-            metas.push(`${key}: ${value.code}`)
+  return [
+    `const ${item.variableName} = `,
+    `() => import(/* webpackChunkName: ${chunkName} */ ${component})`
+  ].join('')
+}
 
-            break
-          }
-        }
+function genRedirect (rule) {
+  const props = []
+
+  props.push(`    path: ${JSON.stringify(rule.from)}`)
+  props.push(`    redirect: ${JSON.stringify(rule.to)}`)
+
+  return `\n  {\n${props.join(',\n')}\n  }`
+}
+
+function genRoute (item) {
+  const props = []
+  const metas = []
+
+  props.push(`    path: ${JSON.stringify(item.path)}`)
+  props.push(`    component: ${item.variableName}`)
+
+  if (item.type === 'dynamic') {
+    const dataPath = pathToFilePath(item.path, 'json')
+    metas.push(`dataPath: ${JSON.stringify(slash(dataPath))}`)
+    metas.push(`dynamic: true`)
+  }
+
+  if (item.meta) {
+    for (const key in item.meta) {
+      const value = item.meta[key]
+
+      if (key[0] === '$') {
+        metas.push(`${key}: ${value}`)
       } else {
         metas.push(`${key}: ${JSON.stringify(value)}`)
       }
     }
+  }
 
-    if (metas.length) {
-      props.push(`    meta: {\n      ${metas.join(',\n      ')}\n    }`)
-    }
+  if (item.name) {
+    props.unshift(`    name: ${JSON.stringify(item.name)}`)
+  }
 
-    if (page.name) {
-      props.unshift(`    name: ${JSON.stringify(page.name)}`)
-    }
+  if (metas.length) {
+    props.push(`    meta: {\n      ${metas.join(',\n      ')}\n    }`)
+  }
 
-    return `\n  {\n${props.join(',\n')}\n  }`
-  }).join(',')}\n]\n\n`
-
-  return res
+  return `\n  {\n${props.join(',\n')}\n  }`
 }
 
-function genChunkName (component, context) {
-  const chunkName = path.relative(context, component)
+function genChunkName (context, route) {
+  const chunkName = relative(context, route.component)
     .split('/')
     .filter(s => s !== '..')
     .map(s => slugify(s))