+++ gridsome/lib/develop.js
@@ -1,84 +1,103 @@
 const fs = require('fs-extra')
 const chalk = require('chalk')
-const { debounce } = require('lodash')
+const isUrl = require('is-url')
+const columnify = require('columnify')
+const resolvePort = require('./server/resolvePort')
+const { prepareUrls, formatPrettyUrl } = require('./server/utils')
+
+const {
+  hasWarnings,
+  logAllWarnings
+} = require('./utils/deprecate')
 
 module.exports = async (context, args) => {
   process.env.NODE_ENV = 'development'
   process.env.GRIDSOME_MODE = 'serve'
 
   const createApp = require('./app')
+  const Server = require('./server/Server')
+
   const app = await createApp(context, { args })
-  const { config } = app
+  const port = await resolvePort(app.config.port)
+  const hostname = app.config.host
+  const urls = prepareUrls(hostname, port)
+  const server = new Server(app, urls)
 
-  const express = require('express')
-  const createExpressServer = require('./server/createExpressServer')
-  const createSockJsServer = require('./server/createSockJsServer')
-  const server = await createExpressServer(app, { withExplorer: true })
-  const sock = await createSockJsServer(app)
-
-  await fs.emptyDir(config.cacheDir)
-
-  server.app.use(config.pathPrefix, express.static(config.staticDir))
-  server.app.use(require('connect-history-api-fallback')())
+  await fs.emptyDir(app.config.cacheDir)
 
   const webpackConfig = await createWebpackConfig(app)
   const compiler = require('webpack')(webpackConfig)
-  server.app.use(require('webpack-hot-middleware')(compiler, {
-    quiet: true,
-    log: false
-  }))
 
-  const devMiddleware = require('webpack-dev-middleware')(compiler, {
-    pathPrefix: webpackConfig.output.pathPrefix,
-    logLevel: 'silent'
+  server.hooks.setup.tap('develop', server => {
+    server.use(require('webpack-hot-middleware')(compiler, {
+      quiet: true,
+      log: false
+    }))
   })
 
-  compiler.hooks.done.tap('gridsome develop', stats => {
+  server.hooks.afterSetup.tap('develop', server => {
+    const devMiddleware = require('webpack-dev-middleware')(compiler, {
+      pathPrefix: webpackConfig.output.pathPrefix,
+      watchOptions: webpackConfig.devServer ? webpackConfig.devServer.watchOptions : null,
+      logLevel: 'silent'
+    })
+
+    server.use(devMiddleware)
+  })
+
+  compiler.hooks.done.tap('develop', stats => {
     if (stats.hasErrors()) {
       return
     }
 
-    console.log()
-    console.log(`  Site running at:          ${chalk.cyan(server.url.site)}`)
-    console.log(`  Explore GraphQL data at:  ${chalk.cyan(server.url.explore)}`)
-    console.log()
-  })
+    const list = []
+    const addTerm = (name, description) => list.push({ name, description })
+    const addSeparator = () => list.push({ name: '' })
 
-  server.app.use((req, res, next) => {
-    return req.originalUrl !== server.endpoint.explore
-      ? devMiddleware(req, res, next)
-      : next()
-  })
-
-  server.app.listen(server.port, server.host, err => {
-    if (err) throw err
-  })
-
-  const createPages = debounce(() => app.createPages(), 16)
-  const fetchQueries = debounce(() => app.broadcast({ type: 'fetch' }), 16)
-  const generateRoutes = debounce(() => app.codegen.generate('routes.js'), 16)
-
-  app.store.on('change', createPages)
-  app.pages.on('create', generateRoutes)
-  app.pages.on('remove', generateRoutes)
-
-  app.pages.on('update', (page, oldPage) => {
-    const { path: oldPath, query: oldQuery } = oldPage
-    const { path, query } = page
-
-    if (
-      (path !== oldPath && !page.internal.isDynamic) ||
-      // pagination was added or removed in page-query
-      (query.paginate && !oldQuery.paginate) ||
-      (!query.paginate && oldQuery.paginate) ||
-      // page-query was created or removed
-      (query.document && !oldQuery.document) ||
-      (!query.document && oldQuery.document)
-    ) {
-      return generateRoutes()
+    if (urls.lan.pretty) {
+      addTerm('Site running at:')
+      addTerm('- Local:', urls.local.pretty)
+      addTerm('- Network:', urls.lan.pretty)
+      addSeparator()
+    } else {
+      addTerm('Site running at:', urls.local.pretty)
     }
 
-    fetchQueries()
+    addTerm('Explore GraphQL data at:', urls.explore.pretty)
+
+    app.compiler.hooks.done.call(
+      { addTerm, addSeparator },
+      { stats, hostname, port, formatPrettyUrl }
+    )
+
+    const columns = list
+      .filter((term, i, terms) => (
+        // remove consecutive separators etc...
+        i && term.name ? terms[i - 1].name !== term.name : true
+      ))
+      .map(term => ({
+        name: term.name,
+        description: isUrl(term.description)
+          ? chalk.cyan(term.description)
+          : term.description
+      }))
+
+    const rendered = columnify(columns, { showHeaders: false })
+
+    console.log()
+    console.log(`  ${rendered.split('\n').join('\n  ')}`)
+    console.log()
+
+    if (hasWarnings()) {
+      console.log()
+      console.log(`${chalk.bgYellow.black(' WARNING ')} ${chalk.yellow('Deprecation notices')}`)
+      console.log()
+      logAllWarnings(app.context)
+    }
+  })
+
+  server.listen(port, hostname, err => {
+    if (err) throw err
   })
 
   //
@@ -86,9 +105,7 @@ module.exports = async (context, args) => {
   //
 
   async function createWebpackConfig (app) {
-    const { SOCKJS_ENDPOINT, GRAPHQL_ENDPOINT, GRAPHQL_WS_ENDPOINT } = process.env
-
-    const config = await app.resolveChainableWebpackConfig()
+    const config = await app.compiler.resolveChainableWebpackConfig()
 
     config
       .plugin('friendly-errors')
@@ -100,19 +117,18 @@ module.exports = async (context, args) => {
         const definitions = args[0]
         args[0] = {
           ...definitions,
-          'process.env.SOCKJS_ENDPOINT': JSON.stringify(SOCKJS_ENDPOINT || sock.url),
-          'process.env.GRAPHQL_ENDPOINT': JSON.stringify(GRAPHQL_ENDPOINT || server.url.graphql),
-          'process.env.GRAPHQL_WS_ENDPOINT': JSON.stringify(GRAPHQL_WS_ENDPOINT || server.url.websocket)
+          'process.env.SOCKJS_ENDPOINT': JSON.stringify(urls.sockjs.endpoint),
+          'process.env.GRAPHQL_ENDPOINT': JSON.stringify(urls.graphql.endpoint)
         }
         return args
       })
 
     config.entryPoints.store.forEach((entry, name) => {
       config.entry(name)
-        .prepend(`webpack-hot-middleware/client?name=${name}&reload=true`)
+        .prepend(`webpack-hot-middleware/client?name=${name}&reload=true&noInfo=true`)
         .prepend('webpack/hot/dev-server')
     })
 
-    return app.resolveWebpackConfig(false, config)
+    return app.compiler.resolveWebpackConfig(false, config)
   }
 }