+++ gridsome/lib/webpack/compileAssets.js
@@ -1,7 +1,7 @@
 module.exports = async (app, defines = {}) => {
   const webpack = require('webpack')
 
-  const clientChain = await app.resolveChainableWebpackConfig()
+  const clientChain = await app.compiler.resolveChainableWebpackConfig()
 
   clientChain
     .plugin('injections')
@@ -15,8 +15,8 @@ module.exports = async (app, defines = {}) => {
     })
 
   const [serverConfig, clientConfig] = await Promise.all([
-    app.resolveWebpackConfig(true),
-    app.resolveWebpackConfig(false, clientChain)
+    app.compiler.resolveWebpackConfig(true),
+    app.compiler.resolveWebpackConfig(false, clientChain)
   ])
 
   return new Promise((resolve, reject) => {
@@ -24,7 +24,12 @@ module.exports = async (app, defines = {}) => {
       if (err) return reject(err)
 
       if (stats.hasErrors()) {
-        const { errors } = stats.toJson()
+        const errors = stats.stats
+          // .flatMap(stats => stats.compilation.errors) only exists in Node v11+
+          .map(stats => stats.compilation.errors)
+          .reduce((acc, errors) => acc.concat(errors), [])
+          .map(err => err.error || err)
+
         return reject(errors[0])
       }