+++ gridsome/lib/webpack/createBaseConfig.js
@@ -5,6 +5,7 @@ const Config = require('webpack-chain')
 const { forwardSlash } = require('../utils')
 const { VueLoaderPlugin } = require('vue-loader')
 const createHTMLRenderer = require('../server/createHTMLRenderer')
+const GridsomeResolverPlugin = require('./plugins/GridsomeResolverPlugin')
 const CSSExtractPlugin = require('mini-css-extract-plugin')
 
 const resolve = (p, c) => path.resolve(c || __dirname, p)
@@ -13,10 +14,10 @@ module.exports = (app, { isProd, isServer }) => {
   const { config: projectConfig } = app
   const { publicPath } = projectConfig
   const { cacheDirectory, cacheIdentifier } = createCacheOptions()
-  const assetsDir = path.relative(projectConfig.outDir, projectConfig.assetsDir)
+  const assetsDir = path.relative(projectConfig.outputDir, projectConfig.assetsDir)
   const config = new Config()
 
-  const useHash = isProd && !process.env.GRIDSOME_TEST
+  const useHash = isProd && !process.env.GRIDSOME_TEST && projectConfig.cacheBusting
   const filename = `[name]${useHash ? '.[contenthash:8]' : ''}.js`
   const assetname = `[name]${useHash ? '.[hash:8]' : ''}.[ext]`
   const inlineLimit = 10000
@@ -25,7 +26,7 @@ module.exports = (app, { isProd, isServer }) => {
 
   config.output
     .publicPath(publicPath)
-    .path(projectConfig.outDir)
+    .path(projectConfig.outputDir)
     .chunkFilename(`${assetsDir}/js/${filename}`)
     .filename(`${assetsDir}/js/${filename}`)
 
@@ -43,6 +44,14 @@ module.exports = (app, { isProd, isServer }) => {
     .add(resolve('../../node_modules'))
     .add(resolve('../../../packages'))
     .add('node_modules')
+
+  config.resolve
+    .plugin('gridsome-fallback-resolver-plugin')
+      .use(GridsomeResolverPlugin, [{
+        fallbackDir: path.join(projectConfig.appPath, 'fallbacks'),
+        optionalDir: path.join(app.context, 'src'),
+        resolve: ['main', 'App.vue']
+      }])
 
   config.resolveLoader
     .set('symlinks', true)
@@ -227,15 +236,19 @@ module.exports = (app, { isProd, isServer }) => {
       }])
   }
 
+  // Short hashes as ids for better long term caching.
+  config.optimization.merge({ moduleIds: 'hashed' })
+
   if (process.env.GRIDSOME_TEST) {
     config.output.pathinfo(true)
     config.optimization.minimize(false)
+    config.optimization.merge({ moduleIds: 'named' })
   }
 
   // helpes
 
   function createCacheOptions () {
-    const values = {
+    const values = app.compiler.hooks.cacheIdentifier.call({
       'gridsome': require('../../package.json').version,
       'cache-loader': require('cache-loader/package.json').version,
       'vue-loader': require('vue-loader/package.json').version,
@@ -245,7 +258,7 @@ module.exports = (app, { isProd, isServer }) => {
       config: (
         (projectConfig.chainWebpack || '').toString()
       )
-    }
+    })
 
     return {
       cacheDirectory: app.resolve('node_modules/.cache/gridsome'),