+++ gridsome/lib/webpack/loaders/assets-loader.js
@@ -3,13 +3,29 @@ const utils = require('loader-utils')
 module.exports = async function (source, map) {
   const callback = this.async()
 
-  const { queue } = process.GRIDSOME
+  const { assets } = process.GRIDSOME
   const options = utils.parseQuery(this.query || '?')
+
+  if (typeof options.imageWidths === 'string') {
+    options.imageWidths = options.imageWidths.split(',').map(Number)
+  }
+  if (typeof options.blur !== 'undefined') {
+    options.blur = parseInt(options.blur, 10)
+  }
+  if (typeof options.width !== 'undefined') {
+    options.width = parseInt(options.width, 10)
+  }
+  if (typeof options.height !== 'undefined') {
+    options.height = parseInt(options.height, 10)
+  }
+  if (typeof options.quality !== 'undefined') {
+    options.quality = parseInt(options.quality, 10)
+  }
 
   let asset
 
   try {
-    asset = await queue.add(this.resourcePath, options)
+    asset = await assets.add(this.resourcePath, options)
   } catch (err) {
     callback(err, source, map)
     return